Datepicker mergestyles conversion phase 1 - file structure
