+++ packages/office-ui-fabric-react/src/components/DatePicker/DatePicker.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
 import { Calendar, ICalendarStrings } from '../Calendar';
-import { DatePicker } from './DatePicker';
+import { DatePickerBase } from './DatePicker.base';
 import { IDatePickerStrings } from './DatePicker.types';
 import { FirstWeekOfYear } from '../../utilities/dateValues/DateValues';
 import { shallow, mount, ReactWrapper } from 'enzyme';
@@ -9,26 +9,26 @@ import { shallow, mount, ReactWrapper } from 'enzyme';
 describe('DatePicker', () => {
   it('renders default DatePicker correctly', () => {
     // This will only render the input. Calendar component has its own snapshot.
-    const component = renderer.create(<DatePicker />);
+    const component = renderer.create(<DatePickerBase />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
   });
 
   it('should not open DatePicker when disabled, no label', () => {
-    const wrapper = mount(<DatePicker disabled />);
+    const wrapper = mount(<DatePickerBase disabled />);
     wrapper.find('i').simulate('click');
 
     expect(wrapper.state('isDatePickerShown')).toBe(false);
   });
 
   it('should not open DatePicker when disabled, with label', () => {
-    const wrapper = mount(<DatePicker disabled label="label" />);
+    const wrapper = mount(<DatePickerBase disabled label="label" />);
     wrapper.find('i').simulate('click');
     expect(wrapper.state('isDatePickerShown')).toBe(false);
   });
 
   describe('when Calendar properties are not specified', () => {
-    const datePicker = shallow(<DatePicker />);
+    const datePicker = shallow(<DatePickerBase />);
     datePicker.setState({ isDatePickerShown: true });
     const calendarProps = datePicker.find(Calendar).props();
 
@@ -68,7 +68,7 @@ describe('DatePicker', () => {
     };
 
     const datePicker = shallow(
-      <DatePicker
+      <DatePickerBase
         isMonthPickerVisible={false}
         showMonthPickerAsOverlay={true}
         value={value}
@@ -155,7 +155,13 @@ describe('DatePicker', () => {
 
     beforeEach(() => {
       datePicker = mount(
-        <DatePicker allowTextInput={true} minDate={minDate} maxDate={maxDate} value={defaultDate} strings={strings} />
+        <DatePickerBase
+          allowTextInput={true}
+          minDate={minDate}
+          maxDate={maxDate}
+          value={defaultDate}
+          strings={strings}
+        />
       );
     });
